auth flow updated
now after the purchase, in supabase "users" table, using the lemonsqueezy webhook, script creates a new row for the user with flag "is_account_created" = 0.
in /auth page it's initially rendering a preform with one "email" input. client sends a request to the server checking whether this email exists in the system (user purchased) and whether this email has an account or not. if not - we render user creation form. otherwise - authorization form